+++ CHANGELOG.md
@@ -3,13 +3,24 @@ All notable changes to this project will be documented in this file.
 `Pageboy` adheres to [Semantic Versioning](http://semver.org/).
 
 #### 1.x Releases
-- `1.3.X` Releases - [1.3.0](#130) | [1.3.1](#131) | [1.3.2](#132) | [1.3.3](#133)
-- `1.2.X` Releases - [1.2.0](#120) | [1.2.1](#121)
-- `1.1.X` Releases - [1.1.0](#110) | [1.1.1](#111) | [1.1.2](#112)
-- `1.0.X` Releases - [1.0.0](#100) | [1.0.1](#101) | [1.0.2](#102) | [1.0.3](#103) | [1.0.4](#104) | [1.0.5](#105) | [1.0.6](#106) | [1.0.7](#107) | [1.0.8](#108) | [1.0.9](#109)
+- `1.4.x` Releases - [1.4.0](#140)
+- `1.3.x` Releases - [1.3.0](#130) | [1.3.1](#131) | [1.3.2](#132) | [1.3.3](#133)
+- `1.2.x` Releases - [1.2.0](#120) | [1.2.1](#121)
+- `1.1.x` Releases - [1.1.0](#110) | [1.1.1](#111) | [1.1.2](#112)
+- `1.0.x` Releases - [1.0.0](#100) | [1.0.1](#101) | [1.0.2](#102) | [1.0.3](#103) | [1.0.4](#104) | [1.0.5](#105) | [1.0.6](#106) | [1.0.7](#107) | [1.0.8](#108) | [1.0.9](#109)
 
 #### 0.x Releases
 - `0.4.X` Releases - [0.4.0](#040) | [0.4.1](#041) | [0.4.2](#042) | [0.4.3](#043) | [0.4.4](#044) | [0.4.5](#045) | [0.4.6](#046) | [0.4.7](#047) | [0.4.8](#048) | [0.4.9](#049) | [0.4.10](#0410) | [0.4.11](#0411) | [0.4.12](#0412)
+
+---
+
+## [1.4.0](https://github.com/uias/Pageboy/releases/tag/1.4.0)
+Released on 2017-07-25
+
+#### Added
+- [#90](https://github.com/uias/Pageboy/issues/90) Support for `UIPageViewController` initialization options.
+     - Added `interPageSpacing` property to `PageboyViewController`.
+     - Added by [msaps](https://github.com/msaps).
 
 ---